lesson_30_1/requirements.txt
psycopg2-binary==2.9.6
pytest==7.4.0
allure-pytest==2.13.2
python-dotenv==1.0.0